rename(#27): SecurityUtils 파일 위치 이동
