+++ review/src/main/java/com/ticketing/review/application/service/ReviewLikeService.java
@@ -3,11 +3,11 @@ package com.ticketing.review.application.service;
 import com.ticketing.review.application.dto.response.ReviewLikeResponseDto;
 import com.ticketing.review.common.exception.ReviewException;
 import com.ticketing.review.common.response.ErrorCode;
-import com.ticketing.review.common.utils.SecurityUtils;
 import com.ticketing.review.domain.model.Review;
 import com.ticketing.review.domain.model.ReviewLike;
 import com.ticketing.review.domain.repository.ReviewLikeRepository;
 import com.ticketing.review.domain.repository.ReviewRepository;
+import com.ticketing.review.infrastructure.utils.SecurityUtils;
 import java.util.Optional;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
