+++ review/src/main/java/com/ticketing/review/application/service/ReviewService.java
@@ -9,9 +9,9 @@ import com.ticketing.review.application.dto.response.ReviewResponseDto;
 import com.ticketing.review.application.dto.response.UpdateReviewResponseDto;
 import com.ticketing.review.common.exception.ReviewException;
 import com.ticketing.review.common.response.ErrorCode;
-import com.ticketing.review.common.utils.SecurityUtils;
 import com.ticketing.review.domain.model.Review;
 import com.ticketing.review.domain.repository.ReviewRepository;
+import com.ticketing.review.infrastructure.utils.SecurityUtils;
 import java.util.List;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
